+++ .storybook/CustomTheme.js
@@ -3,7 +3,6 @@ import logo from "./public/nypl-reservoir-design-system-logo.svg";
 
 export default create({
   base: "light",
-  brandTitle: "NYPL Design System React Library",
-  brandUrl: "https://nypl.org",
+  brandTitle: "Reservoir Design System React Library",
   brandImage: logo,
 });